Implement initial support for Scala 3 and provide semi-automatic derivation for all codecs
