+++ .github/workflows/ci.yml
@@ -22,7 +22,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
-        scala: [2.13.8, 2.12.15]
+        scala: [2.12.15, 2.13.8, 3.1.2]
         java: [temurin@11]
     runs-on: ${{ matrix.os }}
     steps:
