+++ .scalafmt.conf
@@ -19,3 +19,9 @@ rewriteTokens = {
   "→": "->"
   "←": "<-"
 }
+
+fileOverride {
+  "glob:**/scala-3/**" {
+    runner.dialect = scala3
+  }
+}
