+++ build.sbt
@@ -1,9 +1,13 @@
 import ReleaseTransformations._
 
-inThisBuild(
+inThisBuild {
+  val scala212 = "2.12.15"
+  val scala213 = "2.13.8"
+  val scala3   = "3.1.2"
+
   List(
-    scalaVersion                        := "2.13.8",
-    crossScalaVersions                  := Seq("2.13.8", "2.12.15"),
+    scalaVersion                        := scala3,
+    crossScalaVersions                  := Seq(scala212, scala213, scala3),
     githubWorkflowPublishTargetBranches := Seq.empty,
     githubWorkflowBuild += WorkflowStep.Sbt(
       name = Option("Coverage Coveralls"),
@@ -14,7 +18,7 @@ inThisBuild(
       )
     )
   )
-)
+}
 addCommandAlias("coverme", "; clean; coverage; test; coverageReport; coverageAggregate")
 
 lazy val root =
@@ -28,21 +32,29 @@ lazy val root =
         val datastax  = "com.datastax.oss"
         val datastaxV = "4.14.0"
 
-        val zio  = "dev.zio"
-        val zioV = "1.0.14"
+        val zio                   = "dev.zio"
+        val zioV                  = "1.0.14"
+        val magnoliaForScala2     = "com.softwaremill.magnolia1_2" %% "magnolia"      % "1.1.2"
+        val scalaReflectForScala2 = "org.scala-lang"                % "scala-reflect" % scalaVersion.value
+        val magnoliaForScala3     = "com.softwaremill.magnolia1_3" %% "magnolia"      % "1.1.1"
 
-        Seq(
-          datastax                        % "java-driver-core"        % datastaxV,
-          "org.scala-lang.modules"       %% "scala-collection-compat" % "2.7.0",
-          "com.softwaremill.magnolia1_2" %% "magnolia"                % "1.1.2",
-          "org.scala-lang"                % "scala-reflect"           % scalaVersion.value,
-          zio                            %% "zio"                     % zioV,
-          zio                            %% "zio-streams"             % zioV,
-          zio                            %% "zio-test"                % zioV     % Test,
-          zio                            %% "zio-test-sbt"            % zioV     % Test,
-          "com.dimafeng"                 %% "testcontainers-scala"    % "0.40.5" % Test,
-          "com.outr"                     %% "scribe-slf4j"            % "3.8.2"  % Test
-        )
+        val coreDependencies =
+          Seq(
+            datastax                  % "java-driver-core"        % datastaxV,
+            "org.scala-lang.modules" %% "scala-collection-compat" % "2.7.0",
+            zio                      %% "zio"                     % zioV,
+            zio                      %% "zio-streams"             % zioV,
+            zio                      %% "zio-test"                % zioV     % Test,
+            zio                      %% "zio-test-sbt"            % zioV     % Test,
+            "com.dimafeng"           %% "testcontainers-scala"    % "0.40.5" % Test,
+            "com.outr"               %% "scribe-slf4j"            % "3.8.2"  % Test
+          )
+
+        val magnolia =
+          if (scalaVersion.value.startsWith("2")) Seq(magnoliaForScala2, scalaReflectForScala2)
+          else Seq(magnoliaForScala3)
+
+        coreDependencies ++ magnolia
       },
       testFrameworks += new TestFramework("zio.test.sbt.ZTestFramework"),
       Test / fork                 := true,
