+++ project/plugins.sbt
@@ -1,5 +1,5 @@
 addSbtPlugin("com.codecommit"            % "sbt-github-actions" % "0.14.2")
 addSbtPlugin("io.github.davidgregory084" % "sbt-tpolecat"       % "0.2.3")
 addSbtPlugin("com.github.sbt"            % "sbt-release"        % "1.1.0")
-addSbtPlugin("org.scoverage"             % "sbt-scoverage"      % "1.9.3")
+addSbtPlugin("org.scoverage"             % "sbt-scoverage"      % "2.0.0-M4")
 addSbtPlugin("org.scoverage"             % "sbt-coveralls"      % "1.3.2")
