+++ src/main/scala/io/kaizensolutions/virgil/codecs/CqlUdtValueEncoder.scala
@@ -2,8 +2,6 @@ package io.kaizensolutions.virgil.codecs
 
 import com.datastax.oss.driver.api.core.`type`.UserDefinedType
 import com.datastax.oss.driver.api.core.data.UdtValue
-import io.kaizensolutions.virgil.annotations.CqlColumn
-import magnolia1._
 
 /**
  * A [[CqlUdtValueEncoder]] encodes a Scala type `A` as a component of a
@@ -72,21 +70,4 @@ object CqlUdtValueEncoder extends UdtValueEncoderMagnoliaDerivation {
       override def encodeByIndex(structure: UdtValue, index: Int, value: A): UdtValue =
         CqlPrimitiveEncoder.encodePrimitiveByIndex(structure, index, value)(prim)
     }
-}
-
-trait UdtValueEncoderMagnoliaDerivation {
-  type Typeclass[T] = CqlUdtValueEncoder[T]
-
-  def join[T](ctx: CaseClass[CqlUdtValueEncoder, T]): CqlUdtValueEncoder.Object[T] =
-    new CqlUdtValueEncoder.Object[T] {
-      override def encode(structure: UdtValue, value: T): UdtValue =
-        ctx.parameters.foldLeft(structure) { case (acc, param) =>
-          val fieldName  = CqlColumn.extractFieldName(param.annotations).getOrElse(param.label)
-          val fieldValue = param.dereference(value)
-          val encoder    = param.typeclass
-          encoder.encodeByFieldName(acc, fieldName, fieldValue)
-        }
-    }
-
-  implicit def derive[T]: CqlUdtValueEncoder.Object[T] = macro Magnolia.gen[T]
 }
