+++ src/main/scala/io/kaizensolutions/virgil/cql/ValueInCql.scala
@@ -16,21 +16,7 @@ private[virgil] trait ValueInCql {
   override def toString: String =
     value.toString
 }
-object ValueInCql {
+object ValueInCql extends ValueInCqlInstances {
   // Type Refinement
   type WithScalaType[Sc] = ValueInCql { type ScalaType = Sc }
-
-  /**
-   * This implicit conversion automatically captures the value and evidence of
-   * the type's Writer in a cql interpolated string that is necessary to write
-   * data into the Datastax statement
-   */
-  implicit def scalaTypeToValueInCqlInterpolator[Scala](
-    in: Scala
-  )(implicit evidence: CqlRowComponentEncoder[Scala]): ValueInCql.WithScalaType[Scala] =
-    new ValueInCql {
-      type ScalaType = Scala
-      val value: Scala                          = in
-      val writer: CqlRowComponentEncoder[Scala] = evidence
-    }
 }
