+++ src/main/scala/io/kaizensolutions/virgil/internal/QueryType.scala
@@ -5,13 +5,13 @@ import zio.{Chunk, NonEmptyChunk}
 
 sealed private[virgil] trait QueryType
 object QueryType {
-  final private[virgil] case class Select[FromCassandra](
+  final private[virgil] case class Select(
     tableName: String,
     columnNames: NonEmptyChunk[String],
     relations: Chunk[Relation]
   ) extends QueryType
 
-  final private[virgil] case class RawCql[FromCassandra] private (
+  final private[virgil] case class RawCql(
     query: String,
     columns: BindMarkers
   ) extends QueryType
