+++ src/test/scala/io/kaizensolutions/virgil/CollectionsSpec.scala
@@ -1,8 +1,6 @@
 package io.kaizensolutions.virgil
 
-import io.kaizensolutions.virgil.annotations.CqlColumn
-import io.kaizensolutions.virgil.cql._
-import io.kaizensolutions.virgil.dsl._
+import io.kaizensolutions.virgil.CollectionsSpecDatatypes._
 import zio.random.Random
 import zio.test.TestAspect.samples
 import zio.test._
@@ -12,7 +10,7 @@ object CollectionsSpec {
     suite("Collections Specification") {
       testM("Read and write a row containing collections") {
         import SimpleCollectionRow._
-        checkM(gen) { expected =>
+        checkM(simpleCollectionRowGen) { expected =>
           for {
             _         <- insert(expected).execute.runDrain
             result    <- select(expected.id).execute.runCollect
@@ -32,7 +30,7 @@ object CollectionsSpec {
           )
       } + testM("Read and write a row containing nested collections") {
         import NestedCollectionRow._
-        checkM(gen) { expected =>
+        checkM(nestedCollectionRowGen) { expected =>
           for {
             _      <- insert(expected).execute.runDrain
             result <- select(expected.a).execute.runCollect
@@ -40,7 +38,7 @@ object CollectionsSpec {
           } yield assertTrue(actual == expected) && assertTrue(result.length == 1)
         }
       } + testM("Read and write a row that contains an option of collections where the option is None") {
-        checkM(OptionCollectionRow.gen) { popRow =>
+        checkM(optionCollectionRowGen) { popRow =>
           for {
             // Please note that Cassandra does not have the concept of nullable collection data.
             // So if you persist an empty collection wrapped in an Option, you'll get back None
@@ -55,112 +53,22 @@ object CollectionsSpec {
         }
       }
     } @@ samples(10)
-}
 
-final case class SimpleCollectionRow(
-  id: Int,
-  @CqlColumn("map_test") mapTest: Map[Int, String],
-  @CqlColumn("set_test") setTest: Set[Long],
-  @CqlColumn("list_test") listTest: Vector[String]
-)
-object SimpleCollectionRow {
-
-  def insert(in: SimpleCollectionRow): CQL[MutationResult] =
-    InsertBuilder("collectionspec_simplecollectiontable")
-      .value("id", in.id)
-      .value("map_test", in.mapTest)
-      .value("set_test", in.setTest)
-      .value("list_test", in.listTest)
-      .build
-
-  def select(id: Int): CQL[SimpleCollectionRow] =
-    SelectBuilder
-      .from("collectionspec_simplecollectiontable")
-      .column("id")
-      .column("map_test")
-      .column("set_test")
-      .column("list_test")
-      .where("id" === id)
-      .build[SimpleCollectionRow]
-
-  val selectAll: CQL[SimpleCollectionRow] =
-    SelectBuilder
-      .from("collectionspec_simplecollectiontable")
-      .columns("id", "map_test", "set_test", "list_test")
-      .build[SimpleCollectionRow]
-
-  def gen: Gen[Random with Sized, SimpleCollectionRow] =
+  def simpleCollectionRowGen: Gen[Random with Sized, SimpleCollectionRow] =
     for {
       id   <- Gen.int(1, 10000000)
       map  <- Gen.mapOf(key = Gen.anyInt, value = Gen.anyString)
       set  <- Gen.setOf(Gen.anyLong)
       list <- Gen.vectorOf(Gen.anyString)
     } yield SimpleCollectionRow(id, map, set, list)
-}
 
-final case class NestedCollectionRow(
-  a: Int,
-  b: Map[Int, Set[Set[Set[Set[Int]]]]]
-)
-object NestedCollectionRow {
-
-  def select(a: Int): CQL[NestedCollectionRow] =
-    SelectBuilder
-      .from("collectionspec_nestedcollectiontable")
-      .column("a")
-      .column("b")
-      .where("a" === a)
-      .build[NestedCollectionRow]
-
-  def insert(in: NestedCollectionRow): CQL[MutationResult] =
-    InsertBuilder("collectionspec_nestedcollectiontable")
-      .value("a", in.a)
-      .value("b", in.b)
-      .build
-
-  def gen: Gen[Random with Sized, NestedCollectionRow] =
+  def nestedCollectionRowGen: Gen[Random with Sized, NestedCollectionRow] =
     for {
       a <- Gen.int(1, 10000000)
       b <- Gen.mapOf(key = Gen.anyInt, value = Gen.setOf(Gen.setOf(Gen.setOf(Gen.setOf(Gen.anyInt)))))
     } yield NestedCollectionRow(a, b)
-}
 
-final case class OptionCollectionRow(
-  id: Int,
-  @CqlColumn("opt_map_test") m: Option[Map[Int, String]],
-  @CqlColumn("opt_set_test") s: Option[Set[Long]],
-  @CqlColumn("opt_list_test") l: Option[Vector[String]]
-)
-object OptionCollectionRow {
-  private val tableName = "collectionspec_optioncollectiontable"
-  def truncate: CQL[MutationResult] =
-    (cql"TRUNCATE " ++ tableName.asCql).mutation
-
-  def insert(in: OptionCollectionRow): CQL[MutationResult] =
-    InsertBuilder(tableName)
-      .value("id", in.id)
-      .value("opt_map_test", in.m)
-      .value("opt_set_test", in.s)
-      .value("opt_list_test", in.l)
-      .build
-
-  def select(id: Int): CQL[OptionCollectionRow] =
-    SelectBuilder
-      .from(tableName)
-      .column("id")
-      .column("opt_map_test")
-      .column("opt_set_test")
-      .column("opt_list_test")
-      .where("id" === id)
-      .build[OptionCollectionRow]
-
-  val selectAll: CQL[OptionCollectionRow] =
-    SelectBuilder
-      .from(tableName)
-      .columns("id", "opt_map_test", "opt_set_test", "opt_list_test")
-      .build[OptionCollectionRow]
-
-  def gen: Gen[Random with Sized, OptionCollectionRow] =
+  def optionCollectionRowGen: Gen[Random with Sized, OptionCollectionRow] =
     for {
       id   <- Gen.int(1, 10000000)
       map  <- Gen.option(Gen.mapOf(key = Gen.anyInt, value = Gen.anyString))
