+++ src/test/scala/io/kaizensolutions/virgil/DeleteBuilderSpec.scala
@@ -1,17 +1,18 @@
 package io.kaizensolutions.virgil
 
-import io.kaizensolutions.virgil.DeleteBuilderSpec_Person._
+import io.kaizensolutions.virgil.DeleteBuilderSpecDatatypes.DeleteBuilderSpec_Person
+import io.kaizensolutions.virgil.DeleteBuilderSpecDatatypes.DeleteBuilderSpec_Person._
 import io.kaizensolutions.virgil.dsl._
+import zio.Has
 import zio.random.Random
 import zio.test.TestAspect.{samples, sequential}
 import zio.test._
-import zio.{Has, NonEmptyChunk}
 
 object DeleteBuilderSpec {
   def deleteBuilderSpec: Spec[Has[CQLExecutor] with Random with Sized with TestConfig, TestFailure[Any], TestSuccess] =
     suite("Delete Builder Specification") {
       testM("Delete the entire row") {
-        checkM(gen) { person =>
+        checkM(deleteBuilderSpec_PersonGen) { person =>
           for {
             _            <- truncate.execute.runDrain
             _            <- insert(person).execute.runDrain
@@ -23,7 +24,7 @@ object DeleteBuilderSpec {
             assertTrue(afterDelete.isEmpty)
         }
       } + testM("Delete columns in a row without deleting the entire row") {
-        checkM(gen) { person =>
+        checkM(deleteBuilderSpec_PersonGen) { person =>
           for {
             _           <- truncate.execute.runDrain
             _           <- insert(person).execute.runDrain
@@ -41,7 +42,7 @@ object DeleteBuilderSpec {
             assertTrue(afterDelete == DeleteBuilderSpec_Person(id = person.id, name = None, age = None))
         }
       } + testM("Conditionally delete preventing a row from being deleted") {
-        checkM(gen) { person =>
+        checkM(deleteBuilderSpec_PersonGen) { person =>
           for {
             _ <- truncate.execute.runDrain
             _ <- insert(person).execute.runDrain
@@ -56,37 +57,8 @@ object DeleteBuilderSpec {
         }
       }
     } @@ sequential @@ samples(4)
-}
 
-final case class DeleteBuilderSpec_Person(id: Int, name: Option[String], age: Option[Int])
-object DeleteBuilderSpec_Person {
-  val tableName                         = "deletebuilderspec_person"
-  val Id                                = "id"
-  val Name                              = "name"
-  val Age                               = "age"
-  val AllColumns: NonEmptyChunk[String] = NonEmptyChunk(Id, Name, Age)
-
-  def find(id: Int): CQL[DeleteBuilderSpec_Person] =
-    SelectBuilder
-      .from(tableName)
-      .columns(Id, Name, Age)
-      .where(Id === id)
-      .build[DeleteBuilderSpec_Person]
-
-  def insert(in: DeleteBuilderSpec_Person): CQL[MutationResult] =
-    InsertBuilder(tableName)
-      .values(
-        Id   -> in.id,
-        Name -> in.name,
-        Age  -> in.age
-      )
-      .ifNotExists
-      .build
-
-  val truncate: CQL[MutationResult] =
-    CQL.truncate(tableName)
-
-  def gen: Gen[Random with Sized, DeleteBuilderSpec_Person] = for {
+  def deleteBuilderSpec_PersonGen: Gen[Random with Sized, DeleteBuilderSpec_Person] = for {
     id   <- Gen.anyInt
     name <- Gen.anyString
     age  <- Gen.anyInt
