+++ src/test/scala/io/kaizensolutions/virgil/InsertBuilderSpec.scala
@@ -1,8 +1,7 @@
 package io.kaizensolutions.virgil
 
-import io.kaizensolutions.virgil.annotations.CqlColumn
+import io.kaizensolutions.virgil.InsertBuilderSpecDatatypes._
 import io.kaizensolutions.virgil.cql._
-import io.kaizensolutions.virgil.dsl._
 import zio.ZIO
 import zio.duration._
 import zio.random.Random
@@ -14,7 +13,7 @@ object InsertBuilderSpec {
   def insertBuilderSpec =
     suite("Insert Builder Specification") {
       testM("Using TTL and exceeding it will cause the result to not be found") {
-        checkM(InsertBuilderSpecPerson.gen) { person =>
+        checkM(insertBuilderSpecPersonGen) { person =>
           val insert = InsertBuilderSpecPerson
             .insert(person)
             .usingTTL(1.second)
@@ -30,7 +29,7 @@ object InsertBuilderSpec {
           insert *> Live.live(ZIO.sleep(1001.milliseconds)) *> find.map(res => assertTrue(res.isEmpty))
         }
       } + testM("Using a timestamp is enforced") {
-        checkM(InsertBuilderSpecPerson.gen, Gen.long(13370000, 13371337)) { (person, timestamp) =>
+        checkM(insertBuilderSpecPersonGen, Gen.long(13370000, 13371337)) { (person, timestamp) =>
           val insert = InsertBuilderSpecPerson
             .insert(person)
             .usingTimestamp(timestamp)
@@ -40,8 +39,8 @@ object InsertBuilderSpec {
           val truncate = InsertBuilderSpecPerson.truncate.executeMutation
 
           val find =
-            (s"""SELECT writetime(${InsertBuilderSpecPerson.Name}) 
-               FROM ${InsertBuilderSpecPerson.tableName} 
+            (s"""SELECT writetime(${InsertBuilderSpecPerson.Name})
+               FROM ${InsertBuilderSpecPerson.tableName}
                WHERE ${InsertBuilderSpecPerson.Id} = """.asCql ++ cql"${person.id}")
               .query[WriteTimeNameResult]
               .execute
@@ -52,37 +51,11 @@ object InsertBuilderSpec {
         }
       }
     } @@ sequential @@ samples(2)
-}
-final case class InsertBuilderSpecPerson(id: Int, name: String, age: Int)
-object InsertBuilderSpecPerson {
-  val tableName = "insertbuilderspec_person"
-  val Id        = "id"
-  val Name      = "name"
-  val Age       = "age"
 
-  def gen: Gen[Random, InsertBuilderSpecPerson] =
+  def insertBuilderSpecPersonGen: Gen[Random, InsertBuilderSpecPerson] =
     for {
       id   <- Gen.int(1, 10000)
       name <- Gen.stringN(5)(Gen.alphaChar)
       age  <- Gen.int(18, 80)
     } yield InsertBuilderSpecPerson(id, name, age)
-
-  def truncate: CQL[MutationResult] = CQL.truncate(tableName)
-
-  def insert(in: InsertBuilderSpecPerson): InsertBuilder[InsertState.ColumnAdded] =
-    InsertBuilder(tableName)
-      .values(
-        Id   -> in.id,
-        Name -> in.name,
-        Age  -> in.age
-      )
-
-  def find(id: Int): CQL[InsertBuilderSpecPerson] =
-    SelectBuilder
-      .from(tableName)
-      .allColumns
-      .where(Id === id)
-      .build[InsertBuilderSpecPerson]
 }
-
-final case class WriteTimeNameResult(@CqlColumn("writetime(name)") result: Long)
