+++ src/test/scala/io/kaizensolutions/virgil/RelationSpec.scala
@@ -1,7 +1,7 @@
 package io.kaizensolutions.virgil
 
-import io.kaizensolutions.virgil.RelationSpec_Person._
-import io.kaizensolutions.virgil.cql.CqlInterpolatedStringOpsForString
+import io.kaizensolutions.virgil.RelationSpecDatatypes._
+import RelationSpec_Person._
 import io.kaizensolutions.virgil.dsl._
 import zio.Has
 import zio.random.Random
@@ -12,7 +12,7 @@ object RelationSpec {
   def relationSpec: ZSpec[TestConfig with Random with Has[CQLExecutor], Any] =
     suite("Relational Operators Specification") {
       testM("isNull") {
-        checkM(gen) { person =>
+        checkM(relationSpec_PersonGen) { person =>
           val update =
             UpdateBuilder(table)
               .set(Name := person.name)
@@ -30,7 +30,7 @@ object RelationSpec {
             find.map(actual => assertTrue(actual == person))
         }
       } + testM("isNotNull") {
-        checkM(RelationSpec_Person.gen) { person =>
+        checkM(relationSpec_PersonGen) { person =>
           val insert  = RelationSpec_Person.insert(person).execute.runDrain
           val newName = person.name + " " + person.name
           val update =
@@ -51,37 +51,8 @@ object RelationSpec {
         }
       }
     } @@ sequential @@ samples(4)
-}
 
-final case class RelationSpec_Person(
-  id: Int,
-  name: String,
-  age: Int
-)
-object RelationSpec_Person {
-  val Id   = "id"
-  val Name = "name"
-  val Age  = "age"
-
-  val table: String = "relationspec_person"
-
-  val truncate: CQL[MutationResult] = s"TRUNCATE TABLE $table".asCql.mutation
-
-  def insert(in: RelationSpec_Person): CQL[MutationResult] =
-    InsertBuilder(table)
-      .value(Id, in.id)
-      .value(Name, in.name)
-      .value(Age, in.age)
-      .build
-
-  def find(id: Int): CQL[RelationSpec_Person] =
-    SelectBuilder
-      .from(table)
-      .columns(Id, Name, Age)
-      .where(Id === id)
-      .build[RelationSpec_Person]
-
-  def gen: Gen[Random, RelationSpec_Person] =
+  def relationSpec_PersonGen: Gen[Random, RelationSpec_Person] =
     for {
       id   <- Gen.int(1, 1000)
       name <- Gen.stringBounded(2, 4)(Gen.alphaChar)
