+++ src/test/scala/io/kaizensolutions/virgil/UpdateBuilderSpec.scala
@@ -1,5 +1,6 @@
 package io.kaizensolutions.virgil
 
+import io.kaizensolutions.virgil.UpdateBuilderSpecDatatypes.UpdateBuilderSpecPerson
 import io.kaizensolutions.virgil.dsl._
 import zio.random.Random
 import zio.test.TestAspect._
@@ -10,7 +11,7 @@ object UpdateBuilderSpec {
     suite("UpdateBuilder Specification") {
       testM("Performing an update will upsert a row") {
         import UpdateBuilderSpecPerson._
-        checkM(gen) { person =>
+        checkM(updateBuilderSpecPersonGen) { person =>
           val update =
             UpdateBuilder(tableName)
               .set(Name := person.name)
@@ -25,7 +26,7 @@ object UpdateBuilderSpec {
       } +
         testM("Updating a column (using IF EXISTS) that does not exist will have no effect") {
           import UpdateBuilderSpecPerson._
-          checkM(gen) { person =>
+          checkM(updateBuilderSpecPersonGen) { person =>
             val updatedAge = person.age + 2
             val update =
               UpdateBuilder(tableName)
@@ -42,7 +43,7 @@ object UpdateBuilderSpec {
         } +
         testM("Updating a column using if conditions will update if met") {
           import UpdateBuilderSpecPerson._
-          checkM(gen) { person =>
+          checkM(updateBuilderSpecPersonGen) { person =>
             val updatedAge = person.age + 10
             val update =
               UpdateBuilder(tableName)
@@ -61,30 +62,8 @@ object UpdateBuilderSpec {
           }
         }
     } @@ samples(4) @@ sequential @@ nondeterministic
-}
 
-final case class UpdateBuilderSpecPerson(id: Int, name: String, age: Int)
-object UpdateBuilderSpecPerson {
-  val tableName: String = "updatebuilderspec_person"
-  val Id                = "id"
-  val Name              = "name"
-  val Age               = "age"
-
-  def find(id: Int) =
-    SelectBuilder
-      .from(tableName)
-      .columns(Id, Name, Age)
-      .where(Id === id)
-      .build[UpdateBuilderSpecPerson]
-
-  def insert(in: UpdateBuilderSpecPerson): CQL[MutationResult] =
-    InsertBuilder(tableName)
-      .value("id", in.id)
-      .value("name", in.name)
-      .value("age", in.age)
-      .build
-
-  def gen: Gen[Random with Sized, UpdateBuilderSpecPerson] = for {
+  def updateBuilderSpecPersonGen: Gen[Random with Sized, UpdateBuilderSpecPerson] = for {
     id   <- Gen.int(1, 10000)
     name <- Gen.string(Gen.alphaChar)
     age  <- Gen.int(18, 90)
